added ios build

+++ tools/build-customers-app.sh
@@ -7,6 +7,7 @@
 # himateTEST
 if [ -d apps/customers ]; then
     export PACKAGE_DIRS=`pwd`/packages
+
     echo "Building meteor apps..."
     (cd apps/customers && meteor build ../../build/customers/ --server=https://app.himate.org)
 
@@ -14,7 +15,10 @@ if [ -d apps/customers ]; then
     (cd build/customers/android/ && jarsigner -verbose -sigalg SHA1withRSA -digestalg SHA1 -keystore himate.keystore  release-unsigned.apk himate-app)
 
     echo "Zipalign..."
+    if [ -f build/customers/android/production.apk ]; then
+     echo "remove old apk..."
      (cd build/customers/android/ && rm production.apk)
+    fi
     (cd build/customers/android/ &&  $ANDROID_HOME/build-tools/23.0.1/zipalign 4 release-unsigned.apk production.apk)
 fi
